pc - add test for SecurityConfig
This test helps to fix a mutation testing coverage gap in an inner class of SecurityConfig.
We tried to use a pitest exclusion in the pom.xml to avoid having to write this test,
but that approach was not successful.

That would be the preferred approach, so if we can find a way to do that,
we can eliminate this test.